move email scan loader tests
