+++ api-js/src/loaders/email-scan/__tests__/load-dkim-connections-by-domain-id.test.js
@@ -5,16 +5,16 @@ const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { toGlobalId } = require('graphql-relay')
 const { setupI18n } = require('@lingui/core')
 
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
-const { makeMigrations } = require('../../migrations')
-const { cleanseInput } = require('../validators')
+const englishMessages = require('../../../locale/en/messages')
+const frenchMessages = require('../../../locale/fr/messages')
+const { makeMigrations } = require('../../../../migrations')
+const { cleanseInput } = require('../../../validators')
 const {
-  dmarcLoaderConnectionsByDomainId,
-  dmarcLoaderByKey,
-} = require('../loaders')
+  dkimLoaderConnectionsByDomainId,
+  dkimLoaderByKey,
+} = require('../..')
 
-describe('when given the load dmarc connection function', () => {
+describe('when given the load dkim connection function', () => {
   let query, drop, truncate, migrate, collections, user, domain, i18n
 
   const consoleWarnOutput = []
@@ -64,26 +64,26 @@ describe('when given the load dmarc connection function', () => {
   })
 
   describe('given a successful load', () => {
-    let dmarcScan1, dmarcScan2
+    let dkimScan1, dkimScan2
     beforeEach(async () => {
-      dmarcScan1 = await collections.dmarc.save({
+      dkimScan1 = await collections.dkim.save({
         timestamp: '2020-10-02T12:43:39Z',
       })
-      dmarcScan2 = await collections.dmarc.save({
+      dkimScan2 = await collections.dkim.save({
         timestamp: '2020-10-03T12:43:39Z',
       })
-      await collections.domainsDMARC.save({
-        _to: dmarcScan1._id,
+      await collections.domainsDKIM.save({
+        _to: dkimScan1._id,
         _from: domain._id,
       })
-      await collections.domainsDMARC.save({
-        _to: dmarcScan2._id,
+      await collections.domainsDKIM.save({
+        _to: dkimScan2._id,
         _from: domain._id,
       })
     })
     describe('using no cursor', () => {
-      it('returns multiple dmarc scans', async () => {
-        const connectionLoader = dmarcLoaderConnectionsByDomainId(
+      it('returns multiple dkim scans', async () => {
+        const connectionLoader = dkimLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
@@ -94,35 +94,35 @@ describe('when given the load dmarc connection function', () => {
           first: 5,
         }
 
-        const dmarcScans = await connectionLoader({
+        const dkimScans = await connectionLoader({
           domainId: domain._id,
           ...connectionArgs,
         })
 
-        const dkimLoader = dmarcLoaderByKey(query)
-        const expectedDmarcScans = await dkimLoader.loadMany([
-          dmarcScan1._key,
-          dmarcScan2._key,
+        const dkimLoader = dkimLoaderByKey(query)
+        const expectedDkimScans = await dkimLoader.loadMany([
+          dkimScan1._key,
+          dkimScan2._key,
         ])
 
-        expectedDmarcScans[0].id = expectedDmarcScans[0]._key
-        expectedDmarcScans[1].id = expectedDmarcScans[1]._key
+        expectedDkimScans[0].id = expectedDkimScans[0]._key
+        expectedDkimScans[1].id = expectedDkimScans[1]._key
 
-        expectedDmarcScans[0].domainId = domain._id
-        expectedDmarcScans[1].domainId = domain._id
+        expectedDkimScans[0].domainId = domain._id
+        expectedDkimScans[1].domainId = domain._id
 
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('dmarc', expectedDmarcScans[0]._key),
+              cursor: toGlobalId('dkim', expectedDkimScans[0]._key),
               node: {
-                ...expectedDmarcScans[0],
+                ...expectedDkimScans[0],
               },
             },
             {
-              cursor: toGlobalId('dmarc', expectedDmarcScans[1]._key),
+              cursor: toGlobalId('dkim', expectedDkimScans[1]._key),
               node: {
-                ...expectedDmarcScans[1],
+                ...expectedDkimScans[1],
               },
             },
           ],
@@ -130,41 +130,41 @@ describe('when given the load dmarc connection function', () => {
           pageInfo: {
             hasNextPage: false,
             hasPreviousPage: false,
-            startCursor: toGlobalId('dmarc', expectedDmarcScans[0]._key),
-            endCursor: toGlobalId('dmarc', expectedDmarcScans[1]._key),
+            startCursor: toGlobalId('dkim', expectedDkimScans[0]._key),
+            endCursor: toGlobalId('dkim', expectedDkimScans[1]._key),
           },
         }
 
-        expect(dmarcScans).toEqual(expectedStructure)
+        expect(dkimScans).toEqual(expectedStructure)
       })
     })
     describe('using after cursor', () => {
-      it('returns dmarc scan(s) after a given node id', async () => {
-        const connectionLoader = dmarcLoaderConnectionsByDomainId(
+      it('returns dkim scan(s) after a given node id', async () => {
+        const connectionLoader = dkimLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const dkimLoader = dmarcLoaderByKey(query)
-        const expectedDmarcScans = await dkimLoader.loadMany([
-          dmarcScan1._key,
-          dmarcScan2._key,
+        const dkimLoader = dkimLoaderByKey(query)
+        const expectedDkimScans = await dkimLoader.loadMany([
+          dkimScan1._key,
+          dkimScan2._key,
         ])
 
-        expectedDmarcScans[0].id = expectedDmarcScans[0]._key
-        expectedDmarcScans[0].domainId = domain._id
+        expectedDkimScans[0].id = expectedDkimScans[0]._key
+        expectedDkimScans[0].domainId = domain._id
 
-        expectedDmarcScans[1].id = expectedDmarcScans[1]._key
-        expectedDmarcScans[1].domainId = domain._id
+        expectedDkimScans[1].id = expectedDkimScans[1]._key
+        expectedDkimScans[1].domainId = domain._id
 
         const connectionArgs = {
           first: 5,
-          after: toGlobalId('dmarc', expectedDmarcScans[0]._key),
+          after: toGlobalId('dkim', expectedDkimScans[0]._key),
         }
 
-        const dmarcScans = await connectionLoader({
+        const dkimScans = await connectionLoader({
           domainId: domain._id,
           ...connectionArgs,
         })
@@ -172,9 +172,9 @@ describe('when given the load dmarc connection function', () => {
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('dmarc', expectedDmarcScans[1]._key),
+              cursor: toGlobalId('dkim', expectedDkimScans[1]._key),
               node: {
-                ...expectedDmarcScans[1],
+                ...expectedDkimScans[1],
               },
             },
           ],
@@ -182,41 +182,41 @@ describe('when given the load dmarc connection function', () => {
           pageInfo: {
             hasNextPage: false,
             hasPreviousPage: true,
-            startCursor: toGlobalId('dmarc', expectedDmarcScans[1]._key),
-            endCursor: toGlobalId('dmarc', expectedDmarcScans[1]._key),
+            startCursor: toGlobalId('dkim', expectedDkimScans[1]._key),
+            endCursor: toGlobalId('dkim', expectedDkimScans[1]._key),
           },
         }
 
-        expect(dmarcScans).toEqual(expectedStructure)
+        expect(dkimScans).toEqual(expectedStructure)
       })
     })
     describe('using before cursor', () => {
-      it('returns dmarc scan(s) before a given node id', async () => {
-        const connectionLoader = dmarcLoaderConnectionsByDomainId(
+      it('returns dkim scan(s) before a given node id', async () => {
+        const connectionLoader = dkimLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const dkimLoader = dmarcLoaderByKey(query)
-        const expectedDmarcScans = await dkimLoader.loadMany([
-          dmarcScan1._key,
-          dmarcScan2._key,
+        const dkimLoader = dkimLoaderByKey(query)
+        const expectedDkimScans = await dkimLoader.loadMany([
+          dkimScan1._key,
+          dkimScan2._key,
         ])
 
-        expectedDmarcScans[0].id = expectedDmarcScans[0]._key
-        expectedDmarcScans[0].domainId = domain._id
+        expectedDkimScans[0].id = expectedDkimScans[0]._key
+        expectedDkimScans[0].domainId = domain._id
 
-        expectedDmarcScans[1].id = expectedDmarcScans[1]._key
-        expectedDmarcScans[1].domainId = domain._id
+        expectedDkimScans[1].id = expectedDkimScans[1]._key
+        expectedDkimScans[1].domainId = domain._id
 
         const connectionArgs = {
           first: 5,
-          before: toGlobalId('dmarc', expectedDmarcScans[1]._key),
+          before: toGlobalId('dkim', expectedDkimScans[1]._key),
         }
 
-        const dmarcScans = await connectionLoader({
+        const dkimScans = await connectionLoader({
           domainId: domain._id,
           ...connectionArgs,
         })
@@ -224,9 +224,9 @@ describe('when given the load dmarc connection function', () => {
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('dmarc', expectedDmarcScans[0]._key),
+              cursor: toGlobalId('dkim', expectedDkimScans[0]._key),
               node: {
-                ...expectedDmarcScans[0],
+                ...expectedDkimScans[0],
               },
             },
           ],
@@ -234,40 +234,40 @@ describe('when given the load dmarc connection function', () => {
           pageInfo: {
             hasNextPage: true,
             hasPreviousPage: false,
-            startCursor: toGlobalId('dmarc', expectedDmarcScans[0]._key),
-            endCursor: toGlobalId('dmarc', expectedDmarcScans[0]._key),
+            startCursor: toGlobalId('dkim', expectedDkimScans[0]._key),
+            endCursor: toGlobalId('dkim', expectedDkimScans[0]._key),
           },
         }
 
-        expect(dmarcScans).toEqual(expectedStructure)
+        expect(dkimScans).toEqual(expectedStructure)
       })
     })
     describe('using first limit', () => {
       it('returns the first n amount of item(s)', async () => {
-        const connectionLoader = dmarcLoaderConnectionsByDomainId(
+        const connectionLoader = dkimLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const dkimLoader = dmarcLoaderByKey(query)
-        const expectedDmarcScans = await dkimLoader.loadMany([
-          dmarcScan1._key,
-          dmarcScan2._key,
+        const dkimLoader = dkimLoaderByKey(query)
+        const expectedDkimScans = await dkimLoader.loadMany([
+          dkimScan1._key,
+          dkimScan2._key,
         ])
 
-        expectedDmarcScans[0].id = expectedDmarcScans[0]._key
-        expectedDmarcScans[0].domainId = domain._id
+        expectedDkimScans[0].id = expectedDkimScans[0]._key
+        expectedDkimScans[0].domainId = domain._id
 
-        expectedDmarcScans[1].id = expectedDmarcScans[1]._key
-        expectedDmarcScans[1].domainId = domain._id
+        expectedDkimScans[1].id = expectedDkimScans[1]._key
+        expectedDkimScans[1].domainId = domain._id
 
         const connectionArgs = {
           first: 1,
         }
 
-        const dmarcScans = await connectionLoader({
+        const dkimScans = await connectionLoader({
           domainId: domain._id,
           ...connectionArgs,
         })
@@ -275,9 +275,9 @@ describe('when given the load dmarc connection function', () => {
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('dmarc', expectedDmarcScans[0]._key),
+              cursor: toGlobalId('dkim', expectedDkimScans[0]._key),
               node: {
-                ...expectedDmarcScans[0],
+                ...expectedDkimScans[0],
               },
             },
           ],
@@ -285,40 +285,37 @@ describe('when given the load dmarc connection function', () => {
           pageInfo: {
             hasNextPage: true,
             hasPreviousPage: false,
-            startCursor: toGlobalId('dmarc', expectedDmarcScans[0]._key),
-            endCursor: toGlobalId('dmarc', expectedDmarcScans[0]._key),
+            startCursor: toGlobalId('dkim', expectedDkimScans[0]._key),
+            endCursor: toGlobalId('dkim', expectedDkimScans[0]._key),
           },
         }
 
-        expect(dmarcScans).toEqual(expectedStructure)
+        expect(dkimScans).toEqual(expectedStructure)
       })
     })
     describe('using last limit', () => {
       it('returns the last n amount of item(s)', async () => {
-        const connectionLoader = dmarcLoaderConnectionsByDomainId(
+        const connectionLoader = dkimLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const dkimLoader = dmarcLoaderByKey(query)
-        const expectedDmarcScans = await dkimLoader.loadMany([
-          dmarcScan1._key,
-          dmarcScan2._key,
+        const dkimLoader = dkimLoaderByKey(query)
+        const expectedDkimScans = await dkimLoader.loadMany([
+          dkimScan1._key,
+          dkimScan2._key,
         ])
 
-        expectedDmarcScans[0].id = expectedDmarcScans[0]._key
-        expectedDmarcScans[0].domainId = domain._id
-
-        expectedDmarcScans[1].id = expectedDmarcScans[1]._key
-        expectedDmarcScans[1].domainId = domain._id
+        expectedDkimScans[1].id = expectedDkimScans[1]._key
+        expectedDkimScans[1].domainId = domain._id
 
         const connectionArgs = {
           last: 1,
         }
 
-        const dmarcScans = await connectionLoader({
+        const dkimScans = await connectionLoader({
           domainId: domain._id,
           ...connectionArgs,
         })
@@ -326,9 +323,9 @@ describe('when given the load dmarc connection function', () => {
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('dmarc', expectedDmarcScans[1]._key),
+              cursor: toGlobalId('dkim', expectedDkimScans[1]._key),
               node: {
-                ...expectedDmarcScans[1],
+                ...expectedDkimScans[1],
               },
             },
           ],
@@ -336,52 +333,52 @@ describe('when given the load dmarc connection function', () => {
           pageInfo: {
             hasNextPage: false,
             hasPreviousPage: true,
-            startCursor: toGlobalId('dmarc', expectedDmarcScans[1]._key),
-            endCursor: toGlobalId('dmarc', expectedDmarcScans[1]._key),
+            startCursor: toGlobalId('dkim', expectedDkimScans[1]._key),
+            endCursor: toGlobalId('dkim', expectedDkimScans[1]._key),
           },
         }
 
-        expect(dmarcScans).toEqual(expectedStructure)
+        expect(dkimScans).toEqual(expectedStructure)
       })
     })
     describe('use date filters', () => {
-      let dmarcScan3
+      let dkimScan3
       beforeEach(async () => {
-        dmarcScan3 = await collections.dmarc.save({
+        dkimScan3 = await collections.dkim.save({
           timestamp: '2020-10-04T12:43:39Z',
         })
-        await collections.domainsDMARC.save({
-          _to: dmarcScan3._id,
+        await collections.domainsDKIM.save({
+          _to: dkimScan3._id,
           _from: domain._id,
         })
       })
       describe('using start date filter', () => {
         it('returns dkim scans at and after the start date', async () => {
-          const connectionLoader = dmarcLoaderConnectionsByDomainId(
+          const connectionLoader = dkimLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
             i18n,
           )
 
-          const dmarcLoader = dmarcLoaderByKey(query)
-          const expectedDmarcScans = await dmarcLoader.loadMany([
-            dmarcScan2._key,
-            dmarcScan3._key,
+          const dkimLoader = dkimLoaderByKey(query)
+          const expectedDkimScans = await dkimLoader.loadMany([
+            dkimScan2._key,
+            dkimScan3._key,
           ])
 
-          expectedDmarcScans[0].id = expectedDmarcScans[0]._key
-          expectedDmarcScans[0].domainId = domain._id
+          expectedDkimScans[0].id = expectedDkimScans[0]._key
+          expectedDkimScans[0].domainId = domain._id
 
-          expectedDmarcScans[1].id = expectedDmarcScans[1]._key
-          expectedDmarcScans[1].domainId = domain._id
+          expectedDkimScans[1].id = expectedDkimScans[1]._key
+          expectedDkimScans[1].domainId = domain._id
 
           const connectionArgs = {
             first: 5,
             startDate: '2020-10-03',
           }
 
-          const dmarcScans = await connectionLoader({
+          const dkimScans = await connectionLoader({
             domainId: domain._id,
             ...connectionArgs,
           })
@@ -389,15 +386,15 @@ describe('when given the load dmarc connection function', () => {
           const expectedStructure = {
             edges: [
               {
-                cursor: toGlobalId('dmarc', expectedDmarcScans[0]._key),
+                cursor: toGlobalId('dkim', expectedDkimScans[0]._key),
                 node: {
-                  ...expectedDmarcScans[0],
+                  ...expectedDkimScans[0],
                 },
               },
               {
-                cursor: toGlobalId('dmarc', expectedDmarcScans[1]._key),
+                cursor: toGlobalId('dkim', expectedDkimScans[1]._key),
                 node: {
-                  ...expectedDmarcScans[1],
+                  ...expectedDkimScans[1],
                 },
               },
             ],
@@ -405,41 +402,41 @@ describe('when given the load dmarc connection function', () => {
             pageInfo: {
               hasNextPage: false,
               hasPreviousPage: true,
-              startCursor: toGlobalId('dmarc', expectedDmarcScans[0]._key),
-              endCursor: toGlobalId('dmarc', expectedDmarcScans[1]._key),
+              startCursor: toGlobalId('dkim', expectedDkimScans[0]._key),
+              endCursor: toGlobalId('dkim', expectedDkimScans[1]._key),
             },
           }
 
-          expect(dmarcScans).toEqual(expectedStructure)
+          expect(dkimScans).toEqual(expectedStructure)
         })
       })
       describe('using end date filter', () => {
         it('returns dkim scans at and before the end date', async () => {
-          const connectionLoader = dmarcLoaderConnectionsByDomainId(
+          const connectionLoader = dkimLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
             i18n,
           )
 
-          const dmarcLoader = dmarcLoaderByKey(query)
-          const expectedDmarcScans = await dmarcLoader.loadMany([
-            dmarcScan1._key,
-            dmarcScan2._key,
+          const dkimLoader = dkimLoaderByKey(query)
+          const expectedDkimScans = await dkimLoader.loadMany([
+            dkimScan1._key,
+            dkimScan2._key,
           ])
 
-          expectedDmarcScans[0].id = expectedDmarcScans[0]._key
-          expectedDmarcScans[0].domainId = domain._id
+          expectedDkimScans[0].id = expectedDkimScans[0]._key
+          expectedDkimScans[0].domainId = domain._id
 
-          expectedDmarcScans[1].id = expectedDmarcScans[1]._key
-          expectedDmarcScans[1].domainId = domain._id
+          expectedDkimScans[1].id = expectedDkimScans[1]._key
+          expectedDkimScans[1].domainId = domain._id
 
           const connectionArgs = {
             first: 5,
             endDate: '2020-10-03T13:50:00Z',
           }
 
-          const dmarcScans = await connectionLoader({
+          const dkimScans = await connectionLoader({
             domainId: domain._id,
             ...connectionArgs,
           })
@@ -447,15 +444,15 @@ describe('when given the load dmarc connection function', () => {
           const expectedStructure = {
             edges: [
               {
-                cursor: toGlobalId('dmarc', expectedDmarcScans[0]._key),
+                cursor: toGlobalId('dkim', expectedDkimScans[0]._key),
                 node: {
-                  ...expectedDmarcScans[0],
+                  ...expectedDkimScans[0],
                 },
               },
               {
-                cursor: toGlobalId('dmarc', expectedDmarcScans[1]._key),
+                cursor: toGlobalId('dkim', expectedDkimScans[1]._key),
                 node: {
-                  ...expectedDmarcScans[1],
+                  ...expectedDkimScans[1],
                 },
               },
             ],
@@ -463,30 +460,28 @@ describe('when given the load dmarc connection function', () => {
             pageInfo: {
               hasNextPage: true,
               hasPreviousPage: false,
-              startCursor: toGlobalId('dmarc', expectedDmarcScans[0]._key),
-              endCursor: toGlobalId('dmarc', expectedDmarcScans[1]._key),
+              startCursor: toGlobalId('dkim', expectedDkimScans[0]._key),
+              endCursor: toGlobalId('dkim', expectedDkimScans[1]._key),
             },
           }
 
-          expect(dmarcScans).toEqual(expectedStructure)
+          expect(dkimScans).toEqual(expectedStructure)
         })
       })
       describe('using start and end date filters', () => {
         it('returns dkim scan on a specific date', async () => {
-          const connectionLoader = dmarcLoaderConnectionsByDomainId(
+          const connectionLoader = dkimLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
             i18n,
           )
 
-          const dmarcLoader = dmarcLoaderByKey(query)
-          const expectedDmarcScans = await dmarcLoader.loadMany([
-            dmarcScan2._key,
-          ])
+          const dkimLoader = dkimLoaderByKey(query)
+          const expectedDkimScans = await dkimLoader.loadMany([dkimScan2._key])
 
-          expectedDmarcScans[0].id = expectedDmarcScans[0]._key
-          expectedDmarcScans[0].domainId = domain._id
+          expectedDkimScans[0].id = expectedDkimScans[0]._key
+          expectedDkimScans[0].domainId = domain._id
 
           const connectionArgs = {
             first: 5,
@@ -494,7 +489,7 @@ describe('when given the load dmarc connection function', () => {
             endDate: '2020-10-03T23:59:59Z',
           }
 
-          const dmarcScans = await connectionLoader({
+          const dkimScans = await connectionLoader({
             domainId: domain._id,
             ...connectionArgs,
           })
@@ -502,9 +497,9 @@ describe('when given the load dmarc connection function', () => {
           const expectedStructure = {
             edges: [
               {
-                cursor: toGlobalId('dmarc', expectedDmarcScans[0]._key),
+                cursor: toGlobalId('dkim', expectedDkimScans[0]._key),
                 node: {
-                  ...expectedDmarcScans[0],
+                  ...expectedDkimScans[0],
                 },
               },
             ],
@@ -512,19 +507,19 @@ describe('when given the load dmarc connection function', () => {
             pageInfo: {
               hasNextPage: true,
               hasPreviousPage: true,
-              startCursor: toGlobalId('dmarc', expectedDmarcScans[0]._key),
-              endCursor: toGlobalId('dmarc', expectedDmarcScans[0]._key),
+              startCursor: toGlobalId('dkim', expectedDkimScans[0]._key),
+              endCursor: toGlobalId('dkim', expectedDkimScans[0]._key),
             },
           }
 
-          expect(dmarcScans).toEqual(expectedStructure)
+          expect(dkimScans).toEqual(expectedStructure)
         })
       })
     })
-    describe('no dmarc scans are found', () => {
+    describe('no dkim scans are found', () => {
       it('returns an empty structure', async () => {
         await truncate()
-        const connectionLoader = dmarcLoaderConnectionsByDomainId(
+        const connectionLoader = dkimLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
@@ -535,7 +530,7 @@ describe('when given the load dmarc connection function', () => {
           first: 5,
         }
 
-        const dmarcScans = await connectionLoader({
+        const dkimScans = await connectionLoader({
           domainId: domain._id,
           ...connectionArgs,
         })
@@ -551,7 +546,7 @@ describe('when given the load dmarc connection function', () => {
           },
         }
 
-        expect(dmarcScans).toEqual(expectedStructure)
+        expect(dkimScans).toEqual(expectedStructure)
       })
     })
   })
@@ -568,9 +563,9 @@ describe('when given the load dmarc connection function', () => {
       })
     })
     describe('given a unsuccessful load', () => {
-      describe('first and last arguments are not set', () => {
-        it('returns an error message', async () => {
-          const connectionLoader = dmarcLoaderConnectionsByDomainId(
+      describe('limits are not set', () => {
+        it('throws an error', async () => {
+          const connectionLoader = dkimLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
@@ -587,18 +582,19 @@ describe('when given the load dmarc connection function', () => {
           } catch (err) {
             expect(err).toEqual(
               new Error(
-                'You must provide a `first` or `last` value to properly paginate the `dmarc` connection.',
+                'You must provide a `first` or `last` value to properly paginate the `dkim` connection.',
               ),
             )
           }
+
           expect(consoleWarnOutput).toEqual([
-            `User: ${user._key} did not have either \`first\` or \`last\` arguments set for: dmarcLoaderConnectionsByDomainId.`,
+            `User: ${user._key} did not have either \`first\` or \`last\` arguments set for: dkimLoaderConnectionsByDomainId.`,
           ])
         })
       })
-      describe('first and last arguments are set', () => {
+      describe('both limits are set', () => {
         it('returns an error message', async () => {
-          const connectionLoader = dmarcLoaderConnectionsByDomainId(
+          const connectionLoader = dkimLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
@@ -607,7 +603,7 @@ describe('when given the load dmarc connection function', () => {
 
           const connectionArgs = {
             first: 1,
-            last: 2,
+            last: 5,
           }
 
           try {
@@ -618,19 +614,20 @@ describe('when given the load dmarc connection function', () => {
           } catch (err) {
             expect(err).toEqual(
               new Error(
-                'Passing both `first` and `last` to paginate the `dmarc` connection is not supported.',
+                'Passing both `first` and `last` to paginate the `dkim` connection is not supported.',
               ),
             )
           }
+
           expect(consoleWarnOutput).toEqual([
-            `User: ${user._key} attempted to have \`first\` and \`last\` arguments set for: dmarcLoaderConnectionsByDomainId.`,
+            `User: ${user._key} attempted to have \`first\` and \`last\` arguments set for: dkimLoaderConnectionsByDomainId.`,
           ])
         })
       })
       describe('limits are set below minimum', () => {
-        describe('first is set', () => {
+        describe('first limit is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dmarcLoaderConnectionsByDomainId(
+            const connectionLoader = dkimLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -638,7 +635,7 @@ describe('when given the load dmarc connection function', () => {
             )
 
             const connectionArgs = {
-              first: -1,
+              first: -5,
             }
 
             try {
@@ -649,18 +646,19 @@ describe('when given the load dmarc connection function', () => {
             } catch (err) {
               expect(err).toEqual(
                 new Error(
-                  '`first` on the `dmarc` connection cannot be less than zero.',
+                  '`first` on the `dkim` connection cannot be less than zero.',
                 ),
               )
             }
+
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`first\` set below zero for: dmarcLoaderConnectionsByDomainId.`,
+              `User: ${user._key} attempted to have \`first\` set below zero for: dkimLoaderConnectionsByDomainId.`,
             ])
           })
         })
-        describe('last is set', () => {
+        describe('last limit is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dmarcLoaderConnectionsByDomainId(
+            const connectionLoader = dkimLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -668,7 +666,7 @@ describe('when given the load dmarc connection function', () => {
             )
 
             const connectionArgs = {
-              last: -2,
+              last: -5,
             }
 
             try {
@@ -679,20 +677,21 @@ describe('when given the load dmarc connection function', () => {
             } catch (err) {
               expect(err).toEqual(
                 new Error(
-                  '`last` on the `dmarc` connection cannot be less than zero.',
+                  '`last` on the `dkim` connection cannot be less than zero.',
                 ),
               )
             }
+
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`last\` set below zero for: dmarcLoaderConnectionsByDomainId.`,
+              `User: ${user._key} attempted to have \`last\` set below zero for: dkimLoaderConnectionsByDomainId.`,
             ])
           })
         })
       })
       describe('limits are set above maximum', () => {
-        describe('first is set', () => {
+        describe('first limit is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dmarcLoaderConnectionsByDomainId(
+            const connectionLoader = dkimLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -700,7 +699,7 @@ describe('when given the load dmarc connection function', () => {
             )
 
             const connectionArgs = {
-              first: 1000,
+              first: 500,
             }
 
             try {
@@ -711,18 +710,19 @@ describe('when given the load dmarc connection function', () => {
             } catch (err) {
               expect(err).toEqual(
                 new Error(
-                  'Requesting 1000 records on the `dmarc` connection exceeds the `first` limit of 100 records.',
+                  'Requesting 500 records on the `dkim` connection exceeds the `first` limit of 100 records.',
                 ),
               )
             }
+
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`first\` set to 1000 for: dmarcLoaderConnectionsByDomainId.`,
+              `User: ${user._key} attempted to have \`first\` set to 500 for: dkimLoaderConnectionsByDomainId.`,
             ])
           })
         })
-        describe('last is set', () => {
+        describe('last limit is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dmarcLoaderConnectionsByDomainId(
+            const connectionLoader = dkimLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -730,7 +730,7 @@ describe('when given the load dmarc connection function', () => {
             )
 
             const connectionArgs = {
-              last: 200,
+              last: 500,
             }
 
             try {
@@ -741,12 +741,13 @@ describe('when given the load dmarc connection function', () => {
             } catch (err) {
               expect(err).toEqual(
                 new Error(
-                  'Requesting 200 records on the `dmarc` connection exceeds the `last` limit of 100 records.',
+                  'Requesting 500 records on the `dkim` connection exceeds the `last` limit of 100 records.',
                 ),
               )
             }
+
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`last\` set to 200 for: dmarcLoaderConnectionsByDomainId.`,
+              `User: ${user._key} attempted to have \`last\` set to 500 for: dkimLoaderConnectionsByDomainId.`,
             ])
           })
         })
@@ -757,7 +758,7 @@ describe('when given the load dmarc connection function', () => {
             it(`returns an error when first set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = dmarcLoaderConnectionsByDomainId(
+              const connectionLoader = dkimLoaderConnectionsByDomainId(
                 query,
                 user._key,
                 cleanseInput,
@@ -782,7 +783,7 @@ describe('when given the load dmarc connection function', () => {
               expect(consoleWarnOutput).toEqual([
                 `User: ${
                   user._key
-                } attempted to have \`first\` set as a ${typeof invalidInput} for: dmarcLoaderConnectionsByDomainId.`,
+                } attempted to have \`first\` set as a ${typeof invalidInput} for: dkimLoaderConnectionsByDomainId.`,
               ])
             })
           })
@@ -792,7 +793,7 @@ describe('when given the load dmarc connection function', () => {
             it(`returns an error when last set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = dmarcLoaderConnectionsByDomainId(
+              const connectionLoader = dkimLoaderConnectionsByDomainId(
                 query,
                 user._key,
                 cleanseInput,
@@ -817,7 +818,7 @@ describe('when given the load dmarc connection function', () => {
               expect(consoleWarnOutput).toEqual([
                 `User: ${
                   user._key
-                } attempted to have \`last\` set as a ${typeof invalidInput} for: dmarcLoaderConnectionsByDomainId.`,
+                } attempted to have \`last\` set as a ${typeof invalidInput} for: dkimLoaderConnectionsByDomainId.`,
               ])
             })
           })
@@ -830,7 +831,7 @@ describe('when given the load dmarc connection function', () => {
           .fn()
           .mockRejectedValue(new Error('Database Error Occurred.'))
 
-        const connectionLoader = dmarcLoaderConnectionsByDomainId(
+        const connectionLoader = dkimLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
@@ -840,7 +841,6 @@ describe('when given the load dmarc connection function', () => {
         const connectionArgs = {
           first: 5,
         }
-
         try {
           await connectionLoader({
             domainId: domain._id,
@@ -848,12 +848,12 @@ describe('when given the load dmarc connection function', () => {
           })
         } catch (err) {
           expect(err).toEqual(
-            new Error('Unable to load dmarc scans. Please try again.'),
+            new Error('Unable to load dkim scans. Please try again.'),
           )
         }
 
         expect(consoleErrorOutput).toEqual([
-          `Database error occurred while user: ${user._key} was trying to get dmarc information for ${domain._id}, error: Error: Database Error Occurred.`,
+          `Database error occurred while user: ${user._key} was trying to get dkim information for ${domain._id}, error: Error: Database Error Occurred.`,
         ])
       })
     })
@@ -866,7 +866,7 @@ describe('when given the load dmarc connection function', () => {
         }
         const query = jest.fn().mockReturnValueOnce(cursor)
 
-        const connectionLoader = dmarcLoaderConnectionsByDomainId(
+        const connectionLoader = dkimLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
@@ -876,7 +876,6 @@ describe('when given the load dmarc connection function', () => {
         const connectionArgs = {
           first: 5,
         }
-
         try {
           await connectionLoader({
             domainId: domain._id,
@@ -884,12 +883,12 @@ describe('when given the load dmarc connection function', () => {
           })
         } catch (err) {
           expect(err).toEqual(
-            new Error('Unable to load dmarc scans. Please try again.'),
+            new Error('Unable to load dkim scans. Please try again.'),
           )
         }
 
         expect(consoleErrorOutput).toEqual([
-          `Cursor error occurred while user: ${user._key} was trying to get dmarc information for ${domain._id}, error: Error: Cursor Error Occurred.`,
+          `Cursor error occurred while user: ${user._key} was trying to get dkim information for ${domain._id}, error: Error: Cursor Error Occurred.`,
         ])
       })
     })
@@ -907,9 +906,9 @@ describe('when given the load dmarc connection function', () => {
       })
     })
     describe('given a unsuccessful load', () => {
-      describe('first and last arguments are not set', () => {
-        it('returns an error message', async () => {
-          const connectionLoader = dmarcLoaderConnectionsByDomainId(
+      describe('limits are not set', () => {
+        it('throws an error', async () => {
+          const connectionLoader = dkimLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
@@ -926,14 +925,15 @@ describe('when given the load dmarc connection function', () => {
           } catch (err) {
             expect(err).toEqual(new Error('todo'))
           }
+
           expect(consoleWarnOutput).toEqual([
-            `User: ${user._key} did not have either \`first\` or \`last\` arguments set for: dmarcLoaderConnectionsByDomainId.`,
+            `User: ${user._key} did not have either \`first\` or \`last\` arguments set for: dkimLoaderConnectionsByDomainId.`,
           ])
         })
       })
-      describe('first and last arguments are set', () => {
+      describe('both limits are set', () => {
         it('returns an error message', async () => {
-          const connectionLoader = dmarcLoaderConnectionsByDomainId(
+          const connectionLoader = dkimLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
@@ -942,7 +942,7 @@ describe('when given the load dmarc connection function', () => {
 
           const connectionArgs = {
             first: 1,
-            last: 2,
+            last: 5,
           }
 
           try {
@@ -953,15 +953,16 @@ describe('when given the load dmarc connection function', () => {
           } catch (err) {
             expect(err).toEqual(new Error('todo'))
           }
+
           expect(consoleWarnOutput).toEqual([
-            `User: ${user._key} attempted to have \`first\` and \`last\` arguments set for: dmarcLoaderConnectionsByDomainId.`,
+            `User: ${user._key} attempted to have \`first\` and \`last\` arguments set for: dkimLoaderConnectionsByDomainId.`,
           ])
         })
       })
       describe('limits are set below minimum', () => {
-        describe('first is set', () => {
+        describe('first limit is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dmarcLoaderConnectionsByDomainId(
+            const connectionLoader = dkimLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -969,7 +970,7 @@ describe('when given the load dmarc connection function', () => {
             )
 
             const connectionArgs = {
-              first: -1,
+              first: -5,
             }
 
             try {
@@ -980,14 +981,15 @@ describe('when given the load dmarc connection function', () => {
             } catch (err) {
               expect(err).toEqual(new Error('todo'))
             }
+
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`first\` set below zero for: dmarcLoaderConnectionsByDomainId.`,
+              `User: ${user._key} attempted to have \`first\` set below zero for: dkimLoaderConnectionsByDomainId.`,
             ])
           })
         })
-        describe('last is set', () => {
+        describe('last limit is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dmarcLoaderConnectionsByDomainId(
+            const connectionLoader = dkimLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -995,7 +997,7 @@ describe('when given the load dmarc connection function', () => {
             )
 
             const connectionArgs = {
-              last: -2,
+              last: -5,
             }
 
             try {
@@ -1006,16 +1008,17 @@ describe('when given the load dmarc connection function', () => {
             } catch (err) {
               expect(err).toEqual(new Error('todo'))
             }
+
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`last\` set below zero for: dmarcLoaderConnectionsByDomainId.`,
+              `User: ${user._key} attempted to have \`last\` set below zero for: dkimLoaderConnectionsByDomainId.`,
             ])
           })
         })
       })
       describe('limits are set above maximum', () => {
-        describe('first is set', () => {
+        describe('first limit is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dmarcLoaderConnectionsByDomainId(
+            const connectionLoader = dkimLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -1023,7 +1026,7 @@ describe('when given the load dmarc connection function', () => {
             )
 
             const connectionArgs = {
-              first: 1000,
+              first: 500,
             }
 
             try {
@@ -1034,14 +1037,15 @@ describe('when given the load dmarc connection function', () => {
             } catch (err) {
               expect(err).toEqual(new Error('todo'))
             }
+
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`first\` set to 1000 for: dmarcLoaderConnectionsByDomainId.`,
+              `User: ${user._key} attempted to have \`first\` set to 500 for: dkimLoaderConnectionsByDomainId.`,
             ])
           })
         })
-        describe('last is set', () => {
+        describe('last limit is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dmarcLoaderConnectionsByDomainId(
+            const connectionLoader = dkimLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -1049,7 +1053,7 @@ describe('when given the load dmarc connection function', () => {
             )
 
             const connectionArgs = {
-              last: 200,
+              last: 500,
             }
 
             try {
@@ -1060,8 +1064,9 @@ describe('when given the load dmarc connection function', () => {
             } catch (err) {
               expect(err).toEqual(new Error('todo'))
             }
+
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`last\` set to 200 for: dmarcLoaderConnectionsByDomainId.`,
+              `User: ${user._key} attempted to have \`last\` set to 500 for: dkimLoaderConnectionsByDomainId.`,
             ])
           })
         })
@@ -1072,7 +1077,7 @@ describe('when given the load dmarc connection function', () => {
             it(`returns an error when first set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = dmarcLoaderConnectionsByDomainId(
+              const connectionLoader = dkimLoaderConnectionsByDomainId(
                 query,
                 user._key,
                 cleanseInput,
@@ -1093,7 +1098,7 @@ describe('when given the load dmarc connection function', () => {
               expect(consoleWarnOutput).toEqual([
                 `User: ${
                   user._key
-                } attempted to have \`first\` set as a ${typeof invalidInput} for: dmarcLoaderConnectionsByDomainId.`,
+                } attempted to have \`first\` set as a ${typeof invalidInput} for: dkimLoaderConnectionsByDomainId.`,
               ])
             })
           })
@@ -1103,7 +1108,7 @@ describe('when given the load dmarc connection function', () => {
             it(`returns an error when last set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = dmarcLoaderConnectionsByDomainId(
+              const connectionLoader = dkimLoaderConnectionsByDomainId(
                 query,
                 user._key,
                 cleanseInput,
@@ -1124,7 +1129,7 @@ describe('when given the load dmarc connection function', () => {
               expect(consoleWarnOutput).toEqual([
                 `User: ${
                   user._key
-                } attempted to have \`last\` set as a ${typeof invalidInput} for: dmarcLoaderConnectionsByDomainId.`,
+                } attempted to have \`last\` set as a ${typeof invalidInput} for: dkimLoaderConnectionsByDomainId.`,
               ])
             })
           })
@@ -1137,7 +1142,7 @@ describe('when given the load dmarc connection function', () => {
           .fn()
           .mockRejectedValue(new Error('Database Error Occurred.'))
 
-        const connectionLoader = dmarcLoaderConnectionsByDomainId(
+        const connectionLoader = dkimLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
@@ -1147,7 +1152,6 @@ describe('when given the load dmarc connection function', () => {
         const connectionArgs = {
           first: 5,
         }
-
         try {
           await connectionLoader({
             domainId: domain._id,
@@ -1158,7 +1162,7 @@ describe('when given the load dmarc connection function', () => {
         }
 
         expect(consoleErrorOutput).toEqual([
-          `Database error occurred while user: ${user._key} was trying to get dmarc information for ${domain._id}, error: Error: Database Error Occurred.`,
+          `Database error occurred while user: ${user._key} was trying to get dkim information for ${domain._id}, error: Error: Database Error Occurred.`,
         ])
       })
     })
@@ -1171,7 +1175,7 @@ describe('when given the load dmarc connection function', () => {
         }
         const query = jest.fn().mockReturnValueOnce(cursor)
 
-        const connectionLoader = dmarcLoaderConnectionsByDomainId(
+        const connectionLoader = dkimLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
@@ -1181,7 +1185,6 @@ describe('when given the load dmarc connection function', () => {
         const connectionArgs = {
           first: 5,
         }
-
         try {
           await connectionLoader({
             domainId: domain._id,
@@ -1192,7 +1195,7 @@ describe('when given the load dmarc connection function', () => {
         }
 
         expect(consoleErrorOutput).toEqual([
-          `Cursor error occurred while user: ${user._key} was trying to get dmarc information for ${domain._id}, error: Error: Cursor Error Occurred.`,
+          `Cursor error occurred while user: ${user._key} was trying to get dkim information for ${domain._id}, error: Error: Cursor Error Occurred.`,
         ])
       })
     })
