+++ api-js/src/loaders/email-scan/__tests__/load-spf-connections-by-domain-id.test.js
@@ -5,17 +5,14 @@ const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { toGlobalId } = require('graphql-relay')
 const { setupI18n } = require('@lingui/core')
 
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
-const { makeMigrations } = require('../../migrations')
-const { cleanseInput } = require('../validators')
-const {
-  dkimResultsLoaderConnectionByDkimId,
-  dkimResultLoaderByKey,
-} = require('../loaders')
+const englishMessages = require('../../../locale/en/messages')
+const frenchMessages = require('../../../locale/fr/messages')
+const { makeMigrations } = require('../../../../migrations')
+const { cleanseInput } = require('../../../validators')
+const { spfLoaderConnectionsByDomainId, spfLoaderByKey } = require('../..')
 
-describe('when given the load dkim results connection function', () => {
-  let query, drop, truncate, migrate, collections, user, dkimScan, i18n
+describe('when given the load spf connection function', () => {
+  let query, drop, truncate, migrate, collections, user, domain, i18n
 
   const consoleWarnOutput = []
   const mockedWarn = (output) => consoleWarnOutput.push(output)
@@ -53,8 +50,9 @@ describe('when given the load dkim results connection function', () => {
       tfaValidated: false,
       emailValidated: false,
     })
-    dkimScan = await collections.dkim.save({
-      timestamp: '2020-10-02T12:43:39Z',
+    domain = await collections.domains.save({
+      domain: 'test.domain.gc.ca',
+      slug: 'test-domain-gc-ca',
     })
   })
 
@@ -63,24 +61,26 @@ describe('when given the load dkim results connection function', () => {
   })
 
   describe('given a successful load', () => {
-    let dkimResult1, dkimResult2
+    let spfScan1, spfScan2
     beforeEach(async () => {
-      dkimResult1 = await collections.dkimResults.save({})
-      dkimResult2 = await collections.dkimResults.save({})
-
-      await collections.dkimToDkimResults.save({
-        _from: dkimScan._id,
-        _to: dkimResult1._id,
+      spfScan1 = await collections.spf.save({
+        timestamp: '2020-10-02T12:43:39Z',
       })
-      await collections.dkimToDkimResults.save({
-        _from: dkimScan._id,
-        _to: dkimResult2._id,
+      spfScan2 = await collections.spf.save({
+        timestamp: '2020-10-03T12:43:39Z',
+      })
+      await collections.domainsSPF.save({
+        _to: spfScan1._id,
+        _from: domain._id,
+      })
+      await collections.domainsSPF.save({
+        _to: spfScan2._id,
+        _from: domain._id,
       })
     })
-
     describe('using no cursor', () => {
-      it('returns multiple dkim results', async () => {
-        const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+      it('returns multiple spf scans', async () => {
+        const connectionLoader = spfLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
@@ -91,35 +91,35 @@ describe('when given the load dkim results connection function', () => {
           first: 5,
         }
 
-        const dkimResults = await connectionLoader({
-          dkimId: dkimScan._id,
+        const spfScans = await connectionLoader({
+          domainId: domain._id,
           ...connectionArgs,
         })
 
-        const dkimResultLoader = dkimResultLoaderByKey(query)
-        const expectedDkimResults = await dkimResultLoader.loadMany([
-          dkimResult1._key,
-          dkimResult2._key,
+        const spfLoader = spfLoaderByKey(query)
+        const expectedSpfScans = await spfLoader.loadMany([
+          spfScan1._key,
+          spfScan2._key,
         ])
 
-        expectedDkimResults[0].id = expectedDkimResults[0]._key
-        expectedDkimResults[1].id = expectedDkimResults[1]._key
+        expectedSpfScans[0].id = expectedSpfScans[0]._key
+        expectedSpfScans[1].id = expectedSpfScans[1]._key
 
-        expectedDkimResults[0].dkimId = dkimScan._id
-        expectedDkimResults[1].dkimId = dkimScan._id
+        expectedSpfScans[0].domainId = domain._id
+        expectedSpfScans[1].domainId = domain._id
 
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('dkimResult', expectedDkimResults[0]._key),
+              cursor: toGlobalId('spf', expectedSpfScans[0]._key),
               node: {
-                ...expectedDkimResults[0],
+                ...expectedSpfScans[0],
               },
             },
             {
-              cursor: toGlobalId('dkimResult', expectedDkimResults[1]._key),
+              cursor: toGlobalId('spf', expectedSpfScans[1]._key),
               node: {
-                ...expectedDkimResults[1],
+                ...expectedSpfScans[1],
               },
             },
           ],
@@ -127,51 +127,51 @@ describe('when given the load dkim results connection function', () => {
           pageInfo: {
             hasNextPage: false,
             hasPreviousPage: false,
-            startCursor: toGlobalId('dkimResult', expectedDkimResults[0]._key),
-            endCursor: toGlobalId('dkimResult', expectedDkimResults[1]._key),
+            startCursor: toGlobalId('spf', expectedSpfScans[0]._key),
+            endCursor: toGlobalId('spf', expectedSpfScans[1]._key),
           },
         }
 
-        expect(dkimResults).toEqual(expectedStructure)
+        expect(spfScans).toEqual(expectedStructure)
       })
     })
     describe('using after cursor', () => {
-      it('returns dkim result(s) after a given node id', async () => {
-        const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+      it('returns spf scan(s) after a given node id', async () => {
+        const connectionLoader = spfLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const dkimResultLoader = dkimResultLoaderByKey(query)
-        const expectedDkimResults = await dkimResultLoader.loadMany([
-          dkimResult1._key,
-          dkimResult2._key,
+        const spfLoader = spfLoaderByKey(query)
+        const expectedSpfScans = await spfLoader.loadMany([
+          spfScan1._key,
+          spfScan2._key,
         ])
 
-        expectedDkimResults[0].id = expectedDkimResults[0]._key
-        expectedDkimResults[0].dkimId = dkimScan._id
+        expectedSpfScans[0].id = expectedSpfScans[0]._key
+        expectedSpfScans[1].id = expectedSpfScans[1]._key
 
-        expectedDkimResults[1].id = expectedDkimResults[1]._key
-        expectedDkimResults[1].dkimId = dkimScan._id
+        expectedSpfScans[0].domainId = domain._id
+        expectedSpfScans[1].domainId = domain._id
 
         const connectionArgs = {
           first: 5,
-          after: toGlobalId('dkimResult', expectedDkimResults[0]._key),
+          after: toGlobalId('spf', expectedSpfScans[0]._key),
         }
 
-        const dkimResults = await connectionLoader({
-          dkimId: dkimScan._id,
+        const spfScans = await connectionLoader({
+          domainId: domain._id,
           ...connectionArgs,
         })
 
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('dkimResult', expectedDkimResults[1]._key),
+              cursor: toGlobalId('spf', expectedSpfScans[1]._key),
               node: {
-                ...expectedDkimResults[1],
+                ...expectedSpfScans[1],
               },
             },
           ],
@@ -179,51 +179,51 @@ describe('when given the load dkim results connection function', () => {
           pageInfo: {
             hasNextPage: false,
             hasPreviousPage: true,
-            startCursor: toGlobalId('dkimResult', expectedDkimResults[1]._key),
-            endCursor: toGlobalId('dkimResult', expectedDkimResults[1]._key),
+            startCursor: toGlobalId('spf', expectedSpfScans[1]._key),
+            endCursor: toGlobalId('spf', expectedSpfScans[1]._key),
           },
         }
 
-        expect(dkimResults).toEqual(expectedStructure)
+        expect(spfScans).toEqual(expectedStructure)
       })
     })
     describe('using before cursor', () => {
-      it('returns dkim result(s) before a given node id', async () => {
-        const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+      it('returns spf scan(s) before a given node id', async () => {
+        const connectionLoader = spfLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const dkimResultLoader = dkimResultLoaderByKey(query)
-        const expectedDkimResults = await dkimResultLoader.loadMany([
-          dkimResult1._key,
-          dkimResult2._key,
+        const spfLoader = spfLoaderByKey(query)
+        const expectedSpfScans = await spfLoader.loadMany([
+          spfScan1._key,
+          spfScan2._key,
         ])
 
-        expectedDkimResults[0].id = expectedDkimResults[0]._key
-        expectedDkimResults[0].dkimId = dkimScan._id
+        expectedSpfScans[0].id = expectedSpfScans[0]._key
+        expectedSpfScans[1].id = expectedSpfScans[1]._key
 
-        expectedDkimResults[1].id = expectedDkimResults[1]._key
-        expectedDkimResults[1].dkimId = dkimScan._id
+        expectedSpfScans[0].domainId = domain._id
+        expectedSpfScans[1].domainId = domain._id
 
         const connectionArgs = {
           first: 5,
-          before: toGlobalId('dkimResult', expectedDkimResults[1]._key),
+          before: toGlobalId('spf', expectedSpfScans[1]._key),
         }
 
-        const dkimResults = await connectionLoader({
-          dkimId: dkimScan._id,
+        const spfScans = await connectionLoader({
+          domainId: domain._id,
           ...connectionArgs,
         })
 
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('dkimResult', expectedDkimResults[0]._key),
+              cursor: toGlobalId('spf', expectedSpfScans[0]._key),
               node: {
-                ...expectedDkimResults[0],
+                ...expectedSpfScans[0],
               },
             },
           ],
@@ -231,50 +231,50 @@ describe('when given the load dkim results connection function', () => {
           pageInfo: {
             hasNextPage: true,
             hasPreviousPage: false,
-            startCursor: toGlobalId('dkimResult', expectedDkimResults[0]._key),
-            endCursor: toGlobalId('dkimResult', expectedDkimResults[0]._key),
+            startCursor: toGlobalId('spf', expectedSpfScans[0]._key),
+            endCursor: toGlobalId('spf', expectedSpfScans[0]._key),
           },
         }
 
-        expect(dkimResults).toEqual(expectedStructure)
+        expect(spfScans).toEqual(expectedStructure)
       })
     })
     describe('using first limit', () => {
       it('returns the first n amount of item(s)', async () => {
-        const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+        const connectionLoader = spfLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const dkimResultLoader = dkimResultLoaderByKey(query)
-        const expectedDkimResults = await dkimResultLoader.loadMany([
-          dkimResult1._key,
-          dkimResult2._key,
+        const spfLoader = spfLoaderByKey(query)
+        const expectedSpfScans = await spfLoader.loadMany([
+          spfScan1._key,
+          spfScan2._key,
         ])
 
-        expectedDkimResults[0].id = expectedDkimResults[0]._key
-        expectedDkimResults[0].dkimId = dkimScan._id
+        expectedSpfScans[0].id = expectedSpfScans[0]._key
+        expectedSpfScans[1].id = expectedSpfScans[1]._key
 
-        expectedDkimResults[1].id = expectedDkimResults[1]._key
-        expectedDkimResults[1].dkimId = dkimScan._id
+        expectedSpfScans[0].domainId = domain._id
+        expectedSpfScans[1].domainId = domain._id
 
         const connectionArgs = {
           first: 1,
         }
 
-        const dkimResults = await connectionLoader({
-          dkimId: dkimScan._id,
+        const spfScans = await connectionLoader({
+          domainId: domain._id,
           ...connectionArgs,
         })
 
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('dkimResult', expectedDkimResults[0]._key),
+              cursor: toGlobalId('spf', expectedSpfScans[0]._key),
               node: {
-                ...expectedDkimResults[0],
+                ...expectedSpfScans[0],
               },
             },
           ],
@@ -282,50 +282,50 @@ describe('when given the load dkim results connection function', () => {
           pageInfo: {
             hasNextPage: true,
             hasPreviousPage: false,
-            startCursor: toGlobalId('dkimResult', expectedDkimResults[0]._key),
-            endCursor: toGlobalId('dkimResult', expectedDkimResults[0]._key),
+            startCursor: toGlobalId('spf', expectedSpfScans[0]._key),
+            endCursor: toGlobalId('spf', expectedSpfScans[0]._key),
           },
         }
 
-        expect(dkimResults).toEqual(expectedStructure)
+        expect(spfScans).toEqual(expectedStructure)
       })
     })
     describe('using last limit', () => {
       it('returns the last n amount of item(s)', async () => {
-        const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+        const connectionLoader = spfLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
           i18n,
         )
 
-        const dkimResultLoader = dkimResultLoaderByKey(query)
-        const expectedDkimResults = await dkimResultLoader.loadMany([
-          dkimResult1._key,
-          dkimResult2._key,
+        const spfLoader = spfLoaderByKey(query)
+        const expectedSpfScans = await spfLoader.loadMany([
+          spfScan1._key,
+          spfScan2._key,
         ])
 
-        expectedDkimResults[0].id = expectedDkimResults[0]._key
-        expectedDkimResults[0].dkimId = dkimScan._id
+        expectedSpfScans[0].id = expectedSpfScans[0]._key
+        expectedSpfScans[1].id = expectedSpfScans[1]._key
 
-        expectedDkimResults[1].id = expectedDkimResults[1]._key
-        expectedDkimResults[1].dkimId = dkimScan._id
+        expectedSpfScans[0].domainId = domain._id
+        expectedSpfScans[1].domainId = domain._id
 
         const connectionArgs = {
           last: 1,
         }
 
-        const dkimResults = await connectionLoader({
-          dkimId: dkimScan._id,
+        const spfScans = await connectionLoader({
+          domainId: domain._id,
           ...connectionArgs,
         })
 
         const expectedStructure = {
           edges: [
             {
-              cursor: toGlobalId('dkimResult', expectedDkimResults[1]._key),
+              cursor: toGlobalId('spf', expectedSpfScans[1]._key),
               node: {
-                ...expectedDkimResults[1],
+                ...expectedSpfScans[1],
               },
             },
           ],
@@ -333,18 +333,193 @@ describe('when given the load dkim results connection function', () => {
           pageInfo: {
             hasNextPage: false,
             hasPreviousPage: true,
-            startCursor: toGlobalId('dkimResult', expectedDkimResults[1]._key),
-            endCursor: toGlobalId('dkimResult', expectedDkimResults[1]._key),
+            startCursor: toGlobalId('spf', expectedSpfScans[1]._key),
+            endCursor: toGlobalId('spf', expectedSpfScans[1]._key),
           },
         }
 
-        expect(dkimResults).toEqual(expectedStructure)
+        expect(spfScans).toEqual(expectedStructure)
       })
     })
-    describe('no dkim results are found', () => {
+    describe('using date filters', () => {
+      let spfScan3
+      beforeEach(async () => {
+        spfScan3 = await collections.spf.save({
+          timestamp: '2020-10-04T12:43:39Z',
+        })
+        await collections.domainsSPF.save({
+          _to: spfScan3._id,
+          _from: domain._id,
+        })
+      })
+      describe('using start date filter', () => {
+        it('returns spf scans at and after the start date', async () => {
+          const connectionLoader = spfLoaderConnectionsByDomainId(
+            query,
+            user._key,
+            cleanseInput,
+            i18n,
+          )
+
+          const spfLoader = spfLoaderByKey(query)
+          const expectedSpfScans = await spfLoader.loadMany([
+            spfScan2._key,
+            spfScan3._key,
+          ])
+
+          expectedSpfScans[0].id = expectedSpfScans[0]._key
+          expectedSpfScans[1].id = expectedSpfScans[1]._key
+
+          expectedSpfScans[0].domainId = domain._id
+          expectedSpfScans[1].domainId = domain._id
+
+          const connectionArgs = {
+            first: 5,
+            startDate: '2020-10-03',
+          }
+
+          const spfScans = await connectionLoader({
+            domainId: domain._id,
+            ...connectionArgs,
+          })
+
+          const expectedStructure = {
+            edges: [
+              {
+                cursor: toGlobalId('spf', expectedSpfScans[0]._key),
+                node: {
+                  ...expectedSpfScans[0],
+                },
+              },
+              {
+                cursor: toGlobalId('spf', expectedSpfScans[1]._key),
+                node: {
+                  ...expectedSpfScans[1],
+                },
+              },
+            ],
+            totalCount: 3,
+            pageInfo: {
+              hasNextPage: false,
+              hasPreviousPage: true,
+              startCursor: toGlobalId('spf', expectedSpfScans[0]._key),
+              endCursor: toGlobalId('spf', expectedSpfScans[1]._key),
+            },
+          }
+
+          expect(spfScans).toEqual(expectedStructure)
+        })
+      })
+      describe('using end date filter', () => {
+        it('returns spf scans at and before the end date', async () => {
+          const connectionLoader = spfLoaderConnectionsByDomainId(
+            query,
+            user._key,
+            cleanseInput,
+            i18n,
+          )
+
+          const spfLoader = spfLoaderByKey(query)
+          const expectedSpfScans = await spfLoader.loadMany([
+            spfScan1._key,
+            spfScan2._key,
+          ])
+
+          expectedSpfScans[0].id = expectedSpfScans[0]._key
+          expectedSpfScans[1].id = expectedSpfScans[1]._key
+
+          expectedSpfScans[0].domainId = domain._id
+          expectedSpfScans[1].domainId = domain._id
+
+          const connectionArgs = {
+            first: 5,
+            endDate: '2020-10-03T13:50:00Z',
+          }
+
+          const spfScans = await connectionLoader({
+            domainId: domain._id,
+            ...connectionArgs,
+          })
+
+          const expectedStructure = {
+            edges: [
+              {
+                cursor: toGlobalId('spf', expectedSpfScans[0]._key),
+                node: {
+                  ...expectedSpfScans[0],
+                },
+              },
+              {
+                cursor: toGlobalId('spf', expectedSpfScans[1]._key),
+                node: {
+                  ...expectedSpfScans[1],
+                },
+              },
+            ],
+            totalCount: 3,
+            pageInfo: {
+              hasNextPage: true,
+              hasPreviousPage: false,
+              startCursor: toGlobalId('spf', expectedSpfScans[0]._key),
+              endCursor: toGlobalId('spf', expectedSpfScans[1]._key),
+            },
+          }
+
+          expect(spfScans).toEqual(expectedStructure)
+        })
+      })
+      describe('using start and end filters', () => {
+        it('returns a scan on a specific date', async () => {
+          const connectionLoader = spfLoaderConnectionsByDomainId(
+            query,
+            user._key,
+            cleanseInput,
+            i18n,
+          )
+
+          const spfLoader = spfLoaderByKey(query)
+          const expectedSpfScans = await spfLoader.loadMany([spfScan2._key])
+
+          expectedSpfScans[0].id = expectedSpfScans[0]._key
+          expectedSpfScans[0].domainId = domain._id
+
+          const connectionArgs = {
+            first: 5,
+            startDate: '2020-10-03T00:00:00Z',
+            endDate: '2020-10-03T23:59:59Z',
+          }
+
+          const spfScans = await connectionLoader({
+            domainId: domain._id,
+            ...connectionArgs,
+          })
+
+          const expectedStructure = {
+            edges: [
+              {
+                cursor: toGlobalId('spf', expectedSpfScans[0]._key),
+                node: {
+                  ...expectedSpfScans[0],
+                },
+              },
+            ],
+            totalCount: 3,
+            pageInfo: {
+              hasNextPage: true,
+              hasPreviousPage: true,
+              startCursor: toGlobalId('spf', expectedSpfScans[0]._key),
+              endCursor: toGlobalId('spf', expectedSpfScans[0]._key),
+            },
+          }
+
+          expect(spfScans).toEqual(expectedStructure)
+        })
+      })
+    })
+    describe('no spf scans are found', () => {
       it('returns an empty structure', async () => {
         await truncate()
-        const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+        const connectionLoader = spfLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
@@ -355,8 +530,8 @@ describe('when given the load dkim results connection function', () => {
           first: 5,
         }
 
-        const dkimResults = await connectionLoader({
-          dkimId: dkimScan._id,
+        const spfScans = await connectionLoader({
+          domainId: domain._id,
           ...connectionArgs,
         })
 
@@ -371,11 +546,11 @@ describe('when given the load dkim results connection function', () => {
           },
         }
 
-        expect(dkimResults).toEqual(expectedStructure)
+        expect(spfScans).toEqual(expectedStructure)
       })
     })
   })
-  describe('language is set to english', () => {
+  describe('users language is set to english', () => {
     beforeAll(() => {
       i18n = setupI18n({
         language: 'en',
@@ -388,9 +563,9 @@ describe('when given the load dkim results connection function', () => {
       })
     })
     describe('given a unsuccessful load', () => {
-      describe('both limits are not set', () => {
+      describe('first and last arguments are not set', () => {
         it('returns an error message', async () => {
-          const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+          const connectionLoader = spfLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
@@ -401,24 +576,24 @@ describe('when given the load dkim results connection function', () => {
 
           try {
             await connectionLoader({
-              dkimId: dkimScan._id,
+              domainId: domain._id,
               ...connectionArgs,
             })
           } catch (err) {
             expect(err).toEqual(
               new Error(
-                'You must provide a `first` or `last` value to properly paginate the `dkimResults` connection.',
+                'You must provide a `first` or `last` value to properly paginate the `spf` connection.',
               ),
             )
           }
           expect(consoleWarnOutput).toEqual([
-            `User: ${user._key} did not have either \`first\` or \`last\` arguments set for: dkimResultsLoaderConnectionByDkimId.`,
+            `User: ${user._key} did not have either \`first\` or \`last\` arguments set for: spfLoaderConnectionsByDomainId.`,
           ])
         })
       })
-      describe('both limits are set', () => {
+      describe('first and last arguments are set', () => {
         it('returns an error message', async () => {
-          const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+          const connectionLoader = spfLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
@@ -432,25 +607,25 @@ describe('when given the load dkim results connection function', () => {
 
           try {
             await connectionLoader({
-              dkimId: dkimScan._id,
+              domainId: domain._id,
               ...connectionArgs,
             })
           } catch (err) {
             expect(err).toEqual(
               new Error(
-                'Passing both `first` and `last` to paginate the `dkimResults` connection is not supported.',
+                'Passing both `first` and `last` to paginate the `spf` connection is not supported.',
               ),
             )
           }
           expect(consoleWarnOutput).toEqual([
-            `User: ${user._key} attempted to have \`first\` and \`last\` arguments set for: dkimResultsLoaderConnectionByDkimId.`,
+            `User: ${user._key} attempted to have \`first\` and \`last\` arguments set for: spfLoaderConnectionsByDomainId.`,
           ])
         })
       })
-      describe('limits are below minimum', () => {
+      describe('limits are set below minimum', () => {
         describe('first is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+            const connectionLoader = spfLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -463,24 +638,24 @@ describe('when given the load dkim results connection function', () => {
 
             try {
               await connectionLoader({
-                dkimId: dkimScan._id,
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
               expect(err).toEqual(
                 new Error(
-                  '`first` on the `dkimResults` connection cannot be less than zero.',
+                  '`first` on the `spf` connection cannot be less than zero.',
                 ),
               )
             }
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`first\` set below zero for: dkimResultsLoaderConnectionByDkimId.`,
+              `User: ${user._key} attempted to have \`first\` set below zero for: spfLoaderConnectionsByDomainId.`,
             ])
           })
         })
         describe('last is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+            const connectionLoader = spfLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -493,26 +668,26 @@ describe('when given the load dkim results connection function', () => {
 
             try {
               await connectionLoader({
-                dkimId: dkimScan._id,
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
               expect(err).toEqual(
                 new Error(
-                  '`last` on the `dkimResults` connection cannot be less than zero.',
+                  '`last` on the `spf` connection cannot be less than zero.',
                 ),
               )
             }
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`last\` set below zero for: dkimResultsLoaderConnectionByDkimId.`,
+              `User: ${user._key} attempted to have \`last\` set below zero for: spfLoaderConnectionsByDomainId.`,
             ])
           })
         })
       })
-      describe('limits are above maximum', () => {
+      describe('limits are set above maximum', () => {
         describe('first is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+            const connectionLoader = spfLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -525,24 +700,24 @@ describe('when given the load dkim results connection function', () => {
 
             try {
               await connectionLoader({
-                dkimId: dkimScan._id,
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
               expect(err).toEqual(
                 new Error(
-                  'Requesting 1000 records on the `dkimResults` connection exceeds the `first` limit of 100 records.',
+                  'Requesting 1000 records on the `spf` connection exceeds the `first` limit of 100 records.',
                 ),
               )
             }
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`first\` set to 1000 for: dkimResultsLoaderConnectionByDkimId.`,
+              `User: ${user._key} attempted to have \`first\` set to 1000 for: spfLoaderConnectionsByDomainId.`,
             ])
           })
         })
         describe('last is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+            const connectionLoader = spfLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -555,18 +730,18 @@ describe('when given the load dkim results connection function', () => {
 
             try {
               await connectionLoader({
-                dkimId: dkimScan._id,
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
               expect(err).toEqual(
                 new Error(
-                  'Requesting 500 records on the `dkimResults` connection exceeds the `last` limit of 100 records.',
+                  'Requesting 500 records on the `spf` connection exceeds the `last` limit of 100 records.',
                 ),
               )
             }
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`last\` set to 500 for: dkimResultsLoaderConnectionByDkimId.`,
+              `User: ${user._key} attempted to have \`last\` set to 500 for: spfLoaderConnectionsByDomainId.`,
             ])
           })
         })
@@ -577,7 +752,7 @@ describe('when given the load dkim results connection function', () => {
             it(`returns an error when first set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+              const connectionLoader = spfLoaderConnectionsByDomainId(
                 query,
                 user._key,
                 cleanseInput,
@@ -602,7 +777,7 @@ describe('when given the load dkim results connection function', () => {
               expect(consoleWarnOutput).toEqual([
                 `User: ${
                   user._key
-                } attempted to have \`first\` set as a ${typeof invalidInput} for: dkimResultsLoaderConnectionByDkimId.`,
+                } attempted to have \`first\` set as a ${typeof invalidInput} for: spfLoaderConnectionsByDomainId.`,
               ])
             })
           })
@@ -612,7 +787,7 @@ describe('when given the load dkim results connection function', () => {
             it(`returns an error when last set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+              const connectionLoader = spfLoaderConnectionsByDomainId(
                 query,
                 user._key,
                 cleanseInput,
@@ -637,7 +812,7 @@ describe('when given the load dkim results connection function', () => {
               expect(consoleWarnOutput).toEqual([
                 `User: ${
                   user._key
-                } attempted to have \`last\` set as a ${typeof invalidInput} for: dkimResultsLoaderConnectionByDkimId.`,
+                } attempted to have \`last\` set as a ${typeof invalidInput} for: spfLoaderConnectionsByDomainId.`,
               ])
             })
           })
@@ -650,7 +825,7 @@ describe('when given the load dkim results connection function', () => {
           .fn()
           .mockRejectedValue(new Error('Database Error Occurred.'))
 
-        const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+        const connectionLoader = spfLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
@@ -660,19 +835,20 @@ describe('when given the load dkim results connection function', () => {
         const connectionArgs = {
           first: 5,
         }
+
         try {
           await connectionLoader({
-            dkimId: dkimScan._id,
+            domainId: domain._id,
             ...connectionArgs,
           })
         } catch (err) {
           expect(err).toEqual(
-            new Error('Unable to load dkim results. Please try again.'),
+            new Error('Unable to load spf scans. Please try again.'),
           )
         }
 
         expect(consoleErrorOutput).toEqual([
-          `Database error occurred while user: ${user._key} was trying to get dkim result information for ${dkimScan._id}, error: Error: Database Error Occurred.`,
+          `Database error occurred while user: ${user._key} was trying to get spf information for ${domain._id}, error: Error: Database Error Occurred.`,
         ])
       })
     })
@@ -685,7 +861,7 @@ describe('when given the load dkim results connection function', () => {
         }
         const query = jest.fn().mockReturnValueOnce(cursor)
 
-        const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+        const connectionLoader = spfLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
@@ -695,24 +871,25 @@ describe('when given the load dkim results connection function', () => {
         const connectionArgs = {
           first: 5,
         }
+
         try {
           await connectionLoader({
-            dkimId: dkimScan._id,
+            domainId: domain._id,
             ...connectionArgs,
           })
         } catch (err) {
           expect(err).toEqual(
-            new Error('Unable to load dkim results. Please try again.'),
+            new Error('Unable to load spf scans. Please try again.'),
           )
         }
 
         expect(consoleErrorOutput).toEqual([
-          `Cursor error occurred while user: ${user._key} was trying to get dkim result information for ${dkimScan._id}, error: Error: Cursor Error Occurred.`,
+          `Cursor error occurred while user: ${user._key} was trying to get spf information for ${domain._id}, error: Error: Cursor Error Occurred.`,
         ])
       })
     })
   })
-  describe('language is set to french', () => {
+  describe('users language is set to french', () => {
     beforeAll(() => {
       i18n = setupI18n({
         language: 'fr',
@@ -725,9 +902,9 @@ describe('when given the load dkim results connection function', () => {
       })
     })
     describe('given a unsuccessful load', () => {
-      describe('both limits are not set', () => {
+      describe('first and last arguments are not set', () => {
         it('returns an error message', async () => {
-          const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+          const connectionLoader = spfLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
@@ -738,20 +915,20 @@ describe('when given the load dkim results connection function', () => {
 
           try {
             await connectionLoader({
-              dkimId: dkimScan._id,
+              domainId: domain._id,
               ...connectionArgs,
             })
           } catch (err) {
             expect(err).toEqual(new Error('todo'))
           }
           expect(consoleWarnOutput).toEqual([
-            `User: ${user._key} did not have either \`first\` or \`last\` arguments set for: dkimResultsLoaderConnectionByDkimId.`,
+            `User: ${user._key} did not have either \`first\` or \`last\` arguments set for: spfLoaderConnectionsByDomainId.`,
           ])
         })
       })
-      describe('both limits are set', () => {
+      describe('first and last arguments are set', () => {
         it('returns an error message', async () => {
-          const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+          const connectionLoader = spfLoaderConnectionsByDomainId(
             query,
             user._key,
             cleanseInput,
@@ -765,21 +942,21 @@ describe('when given the load dkim results connection function', () => {
 
           try {
             await connectionLoader({
-              dkimId: dkimScan._id,
+              domainId: domain._id,
               ...connectionArgs,
             })
           } catch (err) {
             expect(err).toEqual(new Error('todo'))
           }
           expect(consoleWarnOutput).toEqual([
-            `User: ${user._key} attempted to have \`first\` and \`last\` arguments set for: dkimResultsLoaderConnectionByDkimId.`,
+            `User: ${user._key} attempted to have \`first\` and \`last\` arguments set for: spfLoaderConnectionsByDomainId.`,
           ])
         })
       })
-      describe('limits are below minimum', () => {
+      describe('limits are set below minimum', () => {
         describe('first is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+            const connectionLoader = spfLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -792,20 +969,20 @@ describe('when given the load dkim results connection function', () => {
 
             try {
               await connectionLoader({
-                dkimId: dkimScan._id,
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
               expect(err).toEqual(new Error('todo'))
             }
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`first\` set below zero for: dkimResultsLoaderConnectionByDkimId.`,
+              `User: ${user._key} attempted to have \`first\` set below zero for: spfLoaderConnectionsByDomainId.`,
             ])
           })
         })
         describe('last is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+            const connectionLoader = spfLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -818,22 +995,22 @@ describe('when given the load dkim results connection function', () => {
 
             try {
               await connectionLoader({
-                dkimId: dkimScan._id,
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
               expect(err).toEqual(new Error('todo'))
             }
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`last\` set below zero for: dkimResultsLoaderConnectionByDkimId.`,
+              `User: ${user._key} attempted to have \`last\` set below zero for: spfLoaderConnectionsByDomainId.`,
             ])
           })
         })
       })
-      describe('limits are above maximum', () => {
+      describe('limits are set above maximum', () => {
         describe('first is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+            const connectionLoader = spfLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -846,20 +1023,20 @@ describe('when given the load dkim results connection function', () => {
 
             try {
               await connectionLoader({
-                dkimId: dkimScan._id,
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
               expect(err).toEqual(new Error('todo'))
             }
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`first\` set to 1000 for: dkimResultsLoaderConnectionByDkimId.`,
+              `User: ${user._key} attempted to have \`first\` set to 1000 for: spfLoaderConnectionsByDomainId.`,
             ])
           })
         })
         describe('last is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+            const connectionLoader = spfLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
@@ -872,14 +1049,14 @@ describe('when given the load dkim results connection function', () => {
 
             try {
               await connectionLoader({
-                dkimId: dkimScan._id,
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
               expect(err).toEqual(new Error('todo'))
             }
             expect(consoleWarnOutput).toEqual([
-              `User: ${user._key} attempted to have \`last\` set to 500 for: dkimResultsLoaderConnectionByDkimId.`,
+              `User: ${user._key} attempted to have \`last\` set to 500 for: spfLoaderConnectionsByDomainId.`,
             ])
           })
         })
@@ -890,7 +1067,7 @@ describe('when given the load dkim results connection function', () => {
             it(`returns an error when first set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+              const connectionLoader = spfLoaderConnectionsByDomainId(
                 query,
                 user._key,
                 cleanseInput,
@@ -911,7 +1088,7 @@ describe('when given the load dkim results connection function', () => {
               expect(consoleWarnOutput).toEqual([
                 `User: ${
                   user._key
-                } attempted to have \`first\` set as a ${typeof invalidInput} for: dkimResultsLoaderConnectionByDkimId.`,
+                } attempted to have \`first\` set as a ${typeof invalidInput} for: spfLoaderConnectionsByDomainId.`,
               ])
             })
           })
@@ -921,7 +1098,7 @@ describe('when given the load dkim results connection function', () => {
             it(`returns an error when last set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+              const connectionLoader = spfLoaderConnectionsByDomainId(
                 query,
                 user._key,
                 cleanseInput,
@@ -942,7 +1119,7 @@ describe('when given the load dkim results connection function', () => {
               expect(consoleWarnOutput).toEqual([
                 `User: ${
                   user._key
-                } attempted to have \`last\` set as a ${typeof invalidInput} for: dkimResultsLoaderConnectionByDkimId.`,
+                } attempted to have \`last\` set as a ${typeof invalidInput} for: spfLoaderConnectionsByDomainId.`,
               ])
             })
           })
@@ -955,7 +1132,7 @@ describe('when given the load dkim results connection function', () => {
           .fn()
           .mockRejectedValue(new Error('Database Error Occurred.'))
 
-        const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+        const connectionLoader = spfLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
@@ -965,9 +1142,10 @@ describe('when given the load dkim results connection function', () => {
         const connectionArgs = {
           first: 5,
         }
+
         try {
           await connectionLoader({
-            dkimId: dkimScan._id,
+            domainId: domain._id,
             ...connectionArgs,
           })
         } catch (err) {
@@ -975,7 +1153,7 @@ describe('when given the load dkim results connection function', () => {
         }
 
         expect(consoleErrorOutput).toEqual([
-          `Database error occurred while user: ${user._key} was trying to get dkim result information for ${dkimScan._id}, error: Error: Database Error Occurred.`,
+          `Database error occurred while user: ${user._key} was trying to get spf information for ${domain._id}, error: Error: Database Error Occurred.`,
         ])
       })
     })
@@ -988,7 +1166,7 @@ describe('when given the load dkim results connection function', () => {
         }
         const query = jest.fn().mockReturnValueOnce(cursor)
 
-        const connectionLoader = dkimResultsLoaderConnectionByDkimId(
+        const connectionLoader = spfLoaderConnectionsByDomainId(
           query,
           user._key,
           cleanseInput,
@@ -998,9 +1176,10 @@ describe('when given the load dkim results connection function', () => {
         const connectionArgs = {
           first: 5,
         }
+
         try {
           await connectionLoader({
-            dkimId: dkimScan._id,
+            domainId: domain._id,
             ...connectionArgs,
           })
         } catch (err) {
@@ -1008,7 +1187,7 @@ describe('when given the load dkim results connection function', () => {
         }
 
         expect(consoleErrorOutput).toEqual([
-          `Cursor error occurred while user: ${user._key} was trying to get dkim result information for ${dkimScan._id}, error: Error: Cursor Error Occurred.`,
+          `Cursor error occurred while user: ${user._key} was trying to get spf information for ${domain._id}, error: Error: Cursor Error Occurred.`,
         ])
       })
     })
